+++ src/feature_collection.rs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 use std::convert::TryFrom;
+use std::iter::FromIterator;
 
 use crate::errors::Error;
 use crate::json::{Deserialize, Deserializer, JsonObject, JsonValue, Serialize, Serializer};
@@ -28,8 +29,6 @@ use crate::{util, Bbox, Feature};
 /// Serialization:
 ///
 /// ```
-/// # extern crate geojson;
-/// # fn main() {
 /// use geojson::FeatureCollection;
 /// use geojson::GeoJson;
 ///
@@ -45,7 +44,18 @@ use crate::{util, Bbox, Feature};
 ///     serialized,
 ///     "{\"features\":[],\"type\":\"FeatureCollection\"}"
 /// );
-/// # }
+/// ```
+///
+/// Collect from an iterator:
+///
+/// ```rust
+/// use geojson::{FeatureCollection, Feature, Value};
+///
+/// let fc: FeatureCollection = (0..10).map(|idx| -> Feature {
+///     let c = idx as f64;
+///     Value::Point(vec![1.0 * c, 2.0 * c, 3.0 * c]).into()
+/// }).collect();
+/// assert_eq!(fc.features.len(), 10);
 /// ```
 #[derive(Clone, Debug, PartialEq)]
 pub struct FeatureCollection {
@@ -142,5 +152,99 @@ impl<'de> Deserialize<'de> for FeatureCollection {
         let val = JsonObject::deserialize(deserializer)?;
 
         FeatureCollection::from_json_object(val).map_err(|e| D::Error::custom(e.to_string()))
+    }
+}
+
+/// Create a [`FeatureCollection`] using the [`collect`]
+/// method on an iterator of `Feature`s. If every item
+/// contains a bounding-box of the same dimension, then the
+/// output has a bounding-box of the union of them.
+/// Otherwise, the output will not have a bounding-box.
+///
+/// [`collect`]: std::iter::Iterator::collect
+impl FromIterator<Feature> for FeatureCollection {
+    fn from_iter<T: IntoIterator<Item = Feature>>(iter: T) -> Self {
+        let mut bbox = Some(vec![]);
+
+        let features = iter
+            .into_iter()
+            .inspect(|feat| {
+                // Try to compute the bounding-box
+
+                let (curr_bbox, curr_len) = match &mut bbox {
+                    Some(curr_bbox) => {
+                        let curr_len = curr_bbox.len();
+                        (curr_bbox, curr_len)
+                    }
+                    None => {
+                        // implies we can't compute a
+                        // bounding-box for this collection
+                        return;
+                    }
+                };
+
+                match &feat.bbox {
+                    None => {
+                        bbox = None;
+                    }
+                    Some(fbox) if fbox.len() == 0 || fbox.len() % 2 != 0 => {
+                        bbox = None;
+                    }
+                    Some(fbox) if curr_len == 0 => {
+                        // First iteration: just copy values from fbox
+                        *curr_bbox = fbox.clone();
+                    }
+                    Some(fbox) if curr_len != fbox.len() => {
+                        bbox = None;
+                    }
+                    Some(fbox) => {
+                        // Update bbox by computing min/max
+                        curr_bbox.iter_mut().zip(fbox.iter()).enumerate().for_each(
+                            |(idx, (bc, fc))| {
+                                if idx < curr_len / 2 {
+                                    // These are the min coords
+                                    *bc = fc.min(*bc);
+                                } else {
+                                    *bc = fc.max(*bc);
+                                }
+                            },
+                        );
+                    }
+                };
+            })
+            .collect();
+        FeatureCollection {
+            bbox,
+            features,
+            foreign_members: None,
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::{Feature, FeatureCollection, Value};
+
+    #[test]
+    fn test_fc_from_iterator() {
+        let features: Vec<Feature> = vec![
+            {
+                let mut feat: Feature = Value::Point(vec![0., 0., 0.]).into();
+                feat.bbox = Some(vec![-1., -1., -1., 1., 1., 1.]);
+                feat
+            },
+            {
+                let mut feat: Feature = Value::MultiPoint(vec![
+                    vec![10., 10., 10.],
+                    vec![11., 11., 11.],
+                ]).into();
+                feat.bbox = Some(vec![10., 10., 10., 11., 11., 11.]);
+                feat
+            },
+        ];
+
+        let fc: FeatureCollection = features.into_iter().collect();
+        assert_eq!(fc.features.len(), 2);
+        assert_eq!(fc.bbox, Some(vec![-1., -1., -1., 11., 11., 11.]));
     }
 }